WIP -> Weather Update
-> Réorganisations de fichiers / dossiers / cleanup

 --- Weather ---
Températures : Revisite des tips
Pression de l'air : Fix des bugs concernant le feedback du click, ajustements
Wind : Fix des problèmes visuel du compass,  on ne peut à présent plus déselectionner, une direction sera forcément sélectionnée.
Humidité: Ajustement de la zone clickable du slider, ajustements et refonte du graphique

Implémentation du reset all et de l'envoie de données.
Ajout d'un Scriptable Object de configuration pour gérer l'argent que gagne le joueur lorsqu'il envoie un report ainsi que la précision "SO_Money_Config_Default"

A PLAYTESTER

à faire: coder un Game Rules, qui permet par exemple à la nightwatch d'être active entre telle et telle heure, de même pour la météo ou autre...

+++ Assets/Player/Currency/PlayerCurrency.cs
@@ -4,17 +4,17 @@ namespace LightHouse.Money
 {
     public static class PlayerCurrency
     {
-        public static int Balance { get; private set; }
+        public static float Balance { get; private set; }
 
-        public static event Action<int> OnBalanceChanged;
+        public static event Action<float> OnBalanceChanged;
 
-        public static void Add(int amount)
+        public static void Add(float amount)
         {
             Balance += amount;
             OnBalanceChanged?.Invoke(Balance);
         }
 
-        public static bool TrySpend(int amount)
+        public static bool TrySpend(float amount)
         {
             if (Balance < amount) return false;
             Balance -= amount;
